# postman/openapi.yml
openapi: 3.0.3
info:
  title: Project Activity API
  version: 1.0.0
  description: |-
    API проектной деятельности студентов.

    ## Авторизация
    JWT Bearer: получите `access` через `POST /api/accounts/login/`, затем Authorize → `Bearer <access>`.

    ## Postman
    Импортируйте схему: Import → Link → `http://localhost:8000/api/schema/` (local) или `https://pd.emiit.ru/api/schema/` (prod). Включите синхронизацию с источником (Keep in sync / Update from definition).
paths:
...
  /api/showcase/project-applications/dashboard/:
    get:
      operationId: showcase_project_applications_dashboard_retrieve
      description: GET /api/showcase/project-applications/dashboard/
      summary: Дашборд проектных заявок
      parameters:
      - in: query
        name: application_type
        schema:
          type: string
        description: all | external | internal
      - in: query
        name: days
        schema:
          type: integer
      - in: query
        name: department_id
        schema:
          type: integer
      - in: query
        name: institute_code
        schema:
          type: string
      - in: query
        name: semester_id
        schema:
          type: string
        required: true
      - in: query
        name: status
        schema:
          type: string
        description: approved,rejected,pending,in_progress (через запятую)
      tags:
      - showcase
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token из /api/accounts/login/
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...